+++ java/spark-1.6.0/sources.scala
@@ -23,25 +23,6 @@ object Sources {
     "connection_is_open <- function(sc) {\n" +
     "  UseMethod(\"connection_is_open\")\n" +
     "}\n" +
-    "# nolint start\n" +
-    "# Type mapping from Java to R\n" +
-    "#\n" +
-    "# void -> NULL\n" +
-    "# Int -> integer\n" +
-    "# String -> character\n" +
-    "# Boolean -> logical\n" +
-    "# Float -> double\n" +
-    "# Double -> double\n" +
-    "# Long -> double\n" +
-    "# Array[Byte] -> raw\n" +
-    "# Date -> Date\n" +
-    "# Time -> POSIXct\n" +
-    "#\n" +
-    "# Array[T] -> list()\n" +
-    "# Object -> jobj\n" +
-    "#\n" +
-    "# nolint end\n" +
-    "\n" +
     "readObject <- function(con) {\n" +
     "  # Read type first\n" +
     "  type <- readType(con)\n" +
@@ -74,6 +55,16 @@ object Sources {
     "  string\n" +
     "}\n" +
     "\n" +
+    "readDateArray <- function(con, n = 1) {\n" +
+    "  dates <- list()\n" +
+    "  for (i in 1:n) {\n" +
+    "    string <- readString(con)\n" +
+    "    dates[[i]] <- as.Date(string)\n" +
+    "  }\n" +
+    "\n" +
+    "  do.call(\"c\", dates)\n" +
+    "}\n" +
+    "\n" +
     "readInt <- function(con, n = 1) {\n" +
     "  readBin(con, integer(), n = n, endian = \"big\")\n" +
     "}\n" +
@@ -103,13 +94,16 @@ object Sources {
     "  type <- readType(con)\n" +
     "  len <- readInt(con)\n" +
     "\n" +
-    "  # short-circuit for reading arrays of double, int, logical\n" +
     "  if (type == \"d\") {\n" +
     "    return(readDouble(con, n = len))\n" +
     "  } else if (type == \"i\") {\n" +
     "    return(readInt(con, n = len))\n" +
     "  } else if (type == \"b\") {\n" +
     "    return(readBoolean(con, n = len))\n" +
+    "  } else if (type == \"t\") {\n" +
+    "    return(readTime(con, n = len))\n" +
+    "  } else if (type == \"D\") {\n" +
+    "    return(readDateArray(con, n = len))\n" +
     "  }\n" +
     "\n" +
     "  if (len > 0) {\n" +