+++ tests/testthat/test-invoke.R
@@ -7,34 +7,42 @@ test_that("we can invoke_static with 0 arguments", {
 })
 
 test_that("we can invoke_static with 1 scalar argument", {
-    expect_equal(invoke_static(sc, "sparklyr.Test", "unary_primitive_int",
+    expect_equal(invoke_static(sc, "sparklyr.Test", "unaryPrimitiveInt",
                                ensure_scalar_integer(5)), 25)
 
-    expect_error(invoke_static(sc, "sparklyr.Test", "unary_primitive_int", NULL))
+    expect_error(invoke_static(sc, "sparklyr.Test", "unaryPrimitiveInt", NULL))
 
-    expect_equal(invoke_static(sc, "sparklyr.Test", "unary_Integer",
+    expect_equal(invoke_static(sc, "sparklyr.Test", "unaryInteger",
                                ensure_scalar_integer(5)), 25)
 
-    expect_error(invoke_static(sc, "sparklyr.Test", "unary_Integer", NULL))
+    expect_error(invoke_static(sc, "sparklyr.Test", "unaryInteger", NULL))
 
-    expect_equal(invoke_static(sc, "sparklyr.Test", "unary_nullable_Integer",
+    expect_equal(invoke_static(sc, "sparklyr.Test", "unaryNullableInteger",
                                ensure_scalar_integer(5)), 25)
 
-    expect_equal(invoke_static(sc, "sparklyr.Test", "unary_nullable_Integer", NULL), -1)
+    expect_equal(invoke_static(sc, "sparklyr.Test", "unaryNullableInteger", NULL), -1)
 })
 
 test_that("we can invoke_static with 1 Seq argument", {
-    expect_equal(invoke_static(sc, "sparklyr.Test", "unary_seq", list(3, 4)), 25)
+    expect_equal(invoke_static(sc, "sparklyr.Test", "unarySeq", list(3, 4)), 25)
 
-    expect_equal(invoke_static(sc, "sparklyr.Test", "unary_nullable_seq", list(3, 4)), 25)
+    expect_equal(invoke_static(sc, "sparklyr.Test", "unaryNullableSeq", list(3, 4)), 25)
 })
 
 test_that("we can invoke_static with null Seq argument", {
-    expect_equal(invoke_static(sc, "sparklyr.Test", "unary_nullable_seq", NULL), -1)
+    expect_equal(invoke_static(sc, "sparklyr.Test", "unaryNullableSeq", NULL), -1)
 })
 
 test_that("infer correct overloaded method", {
     expect_equal(invoke_static(sc, "sparklyr.Test", "infer", 0), "Double")
     expect_equal(invoke_static(sc, "sparklyr.Test", "infer", "a"), "String")
     expect_equal(invoke_static(sc, "sparklyr.Test", "infer", list()), "Seq")
+})
+
+test_that("roundtrip date array", {
+  dates <- list(as.Date("2016/1/1"), as.Date("2016/1/1"))
+  expect_equal(
+    invoke_static(sc, "sparklyr.Test", "roundtrip", dates),
+    do.call("c", dates)
+  )
 })